+++ src/Element/Html.php
@@ -148,15 +148,6 @@
     /**
      * @return string
      */
-    public function getAttributesAsString() {
-      trigger_error('Deprecated. use Html::renderAttributes', E_USER_DEPRECATED);
-      return static::renderAttributes($this->attributes);
-    }
-
-
-    /**
-     * @return string
-     */
     public function render() {
       return static::tag($this->tag, $this->attributes, $this->getContent());
     }